+++ src/server.ts
@@ -1,7 +1,7 @@
 import express, { type Express } from "express";
 import helmet from "helmet";
 import { videosRouter } from "./services/videos";
-import { errorHandler } from "./error-handler";
+import { errorHandler } from "./utils/error-handler";
 import { httpLogger, logger } from "./utils/logger";
 
 const app: Express = express();
